Add shared tag system for tagging individual mods
Adds a new system of shared tags that are saved in the Penumbra config, and can
then be 1-click added or removed to/from mods via a popup menu. The use case for
this new system is to allow users to more easily re-use tags and to allow them
to quickly tag individual mods.

Shared tags can be added/removed/modified via a new Tags section of the main
Penumbra Settings tab. Once any shared tags have been saved, they can be added
via a new tags button that shows up in the Description and Edit Mod tabs, to the
right of the existing + button that already existed for typing in new tags.

Shared tags have the same restrictions as regular mod tags, and the application
of shared tags should respect the same limits as application of normal tags.

Signed-off-by: AeAstralis <causal_inverse@fastmail.com>